Default Countdown Timer Functionality
Default timer counts down to 01-01-2022 since the year of launch has
been given (but the date and time was not listed).

Timer counts down in days, hours, minutes, and seconds. Years could be
added reasonably easily, but there could be an issue due to leap year.

Known Issues:
- There may be a memory leak which I have tried to resolve in several
ways which seem to have proven unsuccessful to this point.
- The labels and toolbar are not displaying in the emulator, but they
appear in the display of the xml and are set to visible.