Polished Timeline and FAQ UI
Made the Timeline Activity a ScrollView as the base layout, making
scrolling easier. An overlay text will now appear when a node is
selected in the timeline. This text will disappear when the screen is
pressed.

Removed text file reading from the FAQ code. The text file is still in
the Assets folder. The Timeline still uses text file reading to
initialize its information.

For Alpha Version XML-style reading should be added for the FAQ and
Timeline.