Made FAQ XML File
XML Document for FAQ is located in Assets/TimelineAssets folder.

At the moment there is an issue with the way the XmlPullParser is set up
such that it doesn't detect the special characters (i.e. "\n") in the
text. I've also added the FAQ into strings.xml as a backup until this
issue is sorted out.