Deleted Old Timeline and Reformatted XMLs
Deleted the files for the old timeline (TimelineActivity and
TimePhaseActivity). The only change that should be happening in non
Facts/Timeline files is that they no longer "import TimelineActivity" as
it's deleted.

Additionally, I figured out how to use the XML Resource files for the
Timeline and FAQ. I added facts.xml and timeline.xml to store relevant
information.