Two Clocks Implemented with Years and Months
Needs to be cleaned up and display is very ugly, but basic functionality
implemented for countdown to launch and countdown to arrival at the
asteroid. Implementation needs to be cleaned up and encapsulated and
redundancies removed. Might consider using JodaTime if it isn't too
large to change how calculation is done.