Completed Image Gallery v2
The new version of the (Image) Gallery implements a RecyclerView with a
GridLayoutManager and custom Adapter. RecyclerViews are said to be more
optimized than the old GridView. Following this commit will be testing
and clean-up of code to ensure that the new Gallery is more efficient
than the old Gallery.