Added New Video Classes and Directory
The Video section of the Gallery will work with a RecyclerView (similar
to the Image section). Thumbnails of videos will be displayed. When one
is selected, the video will be shown full screen (a different Activity).

The videosClasses subdirectory was also added in the gallery
subdirectory for better organization.

Video functionality is still not implemented in app.