Prototype Video Player
The Video section of the Gallery will play a video stored in-app when a
thumbnail is selected.

If we wish to implement videos from a website, we'll need to know which
service Psyche will continue to use in order to implement an API.

The added line to the gitignore refers to the IDE's code style settings.