Revised Timeline Functionality
Now that the Psyche website is live, we have an official Timeline to
base ours on. The Psyche Team chose to split their timeline by phase, so
the UI has been changed to have six tabs (Phases A-F) to match it.

Additionally, clicking on a piece of text will now load another activity
with a more detailed description.

Backend-wise, the Timeline was reprogrammed to function like the
Gallery. The UI uses a RecyclerView and custom Adapter (TextViews
instead of ImageViews) instead of a single TextView containing all of
the information. This has also led to the `timeline.xml` file being
restructured. Arrays that are used in the Timeline are actually defined
in the XML doc, allowing for programmers to edit content without going
into hardcoded string arrays.

The timeline XML doc still needs to be updated with the rest of the info
available on the Psyche website.