Countdown and Timeline Cosmetic Changes
Made it so timeline text will change color with the theme. Increased
size and padding of countdown spinner items. Added countdown circles and
labels to a constraint which is now centered and it should adapt better
to larger screens (appear further down the screen). Removed a lot of
unnecessary code from countdown activity. Increased clickable space
around spinner arrow (but not the size of the arrow icon).